refactor(skills): move skill freshness logic to a new module
