+++ apps/web/src/pages/profiles/profile-skills-section.tsx
@@ -3,8 +3,8 @@ import type {
   SkillSummary,
   SkillUsageSummary,
 } from "@nakama/core/contract";
-import { SKILL_STALE_AFTER_MS } from "@nakama/core/skills/archive";
 import { BUNDLED_SKILL_NAMES } from "@nakama/core/skills/bundled-names";
+import { SKILL_STALE_AFTER_MS } from "@nakama/core/skills/freshness";
 import { BASH_TOOL_ID } from "@nakama/core/tools/protected";
 import { Delete02Icon } from "hugeicons-react";
 import { useMemo } from "react";
