+++ packages/core/package.json
@@ -40,6 +40,7 @@
     "./whatsapp-config": "./src/whatsapp-config.ts",
     "./whatsapp-worker": "./src/whatsapp-worker.ts",
     "./skills/archive": "./src/skills/archive.ts",
+    "./skills/freshness": "./src/skills/freshness.ts",
     "./skills/bundled-names": "./src/skills/bundled-names.ts",
     "./skills/post-turn-review": "./src/skills/post-turn-review.ts",
     "./skills/write": "./src/skills/write.ts",
